+++ src/index.html
@@ -3,7 +3,7 @@
 
 <head>
   <meta charset="utf-8" />
-  <title>Today's To Do List</title>
+  <title>Leaderboard</title>
   <script src="https://kit.fontawesome.com/7622500898.js" crossorigin="anonymous"></script>
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
 </head>
